requirements.txt
robotframework
robotframework-seleniumlibrary
selenium
webdriver-manager
requests
robotframework-jsonlibrary
JSONLibrary